first commit
This image can be used for deployment on a raspberry pi.
It is created to read the enecsys gateway and inverters. and output the
data to pvoutput.
you need at least a 16gb sd/micro sd card.
it has 2 scripts included which can be started after launching the rpi.
1. staticip.sh = to be used to change the ipadrress to a static ip.
2. e2pv_install_update.sh = to download the latest files of omoerbeek
his script and start te configuration.
link to omoerbeek his script: https://github.com/omoerbeek/e2pv